Add more support for spellIds in buff names
SpellIds are now also supported without a full scan.
Since this still uses UnitAura mutliple spells with the same name
will trip up without full scans. On importing triggers with spellIds
the spellName is reset, so that sharing auras with spellIds works
across languages.

Note: We might want to reset the spellname also if the client
language changed, (If we can detect that.) but this patch doesn't
do that.