Allow a delay in an AuraBar
This will cause the bar to ignore a percentage of its timer;  For example, if a skill has a 10 second cooldown and the delay is set to .5, the bar will remain at 100% until 5 seconds have passed, and then will tick down from 100% to 0 between the 5 and 10 second mark.

This is intended to be used with the Remaining Time trigger in order to display a percentage of a timer as if it were the full timer.