Add WeakAuras.GetNameAndIcon
But only implement it partially. For the generic trigger, having a
implementation, makes it easier to make a few things local to
GenericTrigger.lua in a future patch.

The buff trigger code is harder to abstract,
and since I don't plan on similarly reworking WeakAurasOptions.lua,
I leave it as is. Event though this looks somewhat strange.

No logic change in this patch.