Add DBM and BigWigs Trigger
A event trigger for announces and a status trigger for timers.

The dbm api is rather limited, so there's no automatic icon possible.
Also I'm not sure I handle everything correctly, but this works with
the test bars.

The BigWigs api has more finesse, so we have a automatic icon.
Also there seems to be a priority parameter on the timers, which
I'm ignoring for now.