Refactor CreateTalentCache()
Thanks to the patience and help of the great people in #wowace (especially nebula) for making this happen. Most of this code existed because you used to query for talents of other classes, since 6.0 you can't do that anymore though. We noticed that db.talent_cache in the SV is actually never used anywhere but could, in the future, maybe hold a list of icon paths of all talents across specs so you always have icons for those up. I still removed the code for it now, in case we never do it.