DBM Trigger: use new callback api
Gets us a icon and adds another potentially useful condition in
the spellId.

The api also provides us with "colorId" and "timerType", which I
have choosen to ignore for now. Both are somewhat useful, e.g.
showing a timer that matches the colorId "phase", would be
useful to some

But I'm keeping it simple for now and this can be easily extended
at a later date.