Fix duplicate moving the list around
WeakAuras.NewDisplayButton() in addition to adding a new button
also scrolled to that button and started the renaming.

Since in the duplicate code path we first call NewDisplayButton and
then add the new aura to the right group, we first scrolled to
a mostly random position.

Fix that, by removing the code to pick and edit a button from
NewDisplayButton and move that to a separate new function
WeakAuras.PickAndEditDisplay.

Call that function at the end of duplicate.

It might be necessary to add a call to WeakAuras.PickDisplay or
WeakAuras.PickAndEditDisplay from other code paths, I checked
the common ones like New/Import/Add to group and remove from
group and they seemed to work.

Ticket-number: 404