Rewrite visibility checking
In Expand() and Collapse() set a flag region.toShow, which indicates
whether the region should be visible. Subsequent calls to Expand() or
Collapse() that don't change toShow just return without performing any
actions. (That fixes 402). While at that remove toHide as it is the
inverse of toShow. Also introduce a region.animatingHide, indicating
that a aura is in its hide animation.

In the dynamic group simply check for region.toShow || region.animateHide.
Since that removes the isVisible() check, this fixes Ticket 453.

Also since auras that animate it's hiding are still positioned as
visible auras, this fixes 156.

Somewhere in this rewrite 454 is fixed too.