Circular Progress: Full Progress should be start-end visible
So, this goes back to the very first version I wrote. Full progress
should be that the progress angle is equal to end angle.

I confussed myself by mainly testing with buff triggers. Those go
from full progress to no progress, thus confusingly but correctly
they must from full to empty and move in the opposite direction.

That is a clockwise buff trigger must move anti-clockwise.

So that a resource trigger going from empty to full moves in
clokcwise direction.

This changes behaviour, but that's better in the long run.

Ticket-number: 486