Fix Full Scan auras
The removeSpellNames patch used as a logic that a trigger.spellId
always trumps a name. For that to work, the full scan auras need
to reset the spellId.

Apparently in the past importing full scan auras was broken,
since this is now done always changing full scan auras broke.

Ticket-number: 516