Clean up region:OnShow handlers
Since the options dialog does not set the OnShow or OnHide handlers,
it shouldn't reset them to nil.

While trying to prove that this causes bugs, I found out that the
existing OnShow handler seems to not have any purpose.

The original intent behind is iirc the following case:
- A aura shows a buff trigger
- A movie is played, which hides all regions
- The buff ends, the region should be hidden
- The movie ends, reshowing all regions that were hidden in
  the second step.

I can't reproduce that case anymore without a OnShow handler.