Fix icon cooldowns having the wrong frame level
If multiple icons are in a group, their frame level are adjusted so
that they each get a unique frame level.

For icons we need to also adjust the frame levels of the cooldown
and stacks frame.

See https://youtu.be/51l5gqQIi7k for what it looks like before and
after.

This seems to be new in Legion.