Fix Ascendance + Lava Burst Cooldown trigger
If Lava Burst is on cooldown and Ascendance is triggered, which
resets the cooldown of LavaBurst the following information is returned
by GetSpellCharges:
StartTime: the old expiration time
Charges: 0, MaxCharges: 2
Duration: 0

So apparently a duration 0 spell has to be treated as if it's not
on cooldown.

Override charges to return 1 in that case. As the Spell Cooldown
trigger has a condition on Charges and I think it's saner if
checking for charges >= 1 shows the aura even if charges is
apparently 0 in that case.

Ticket-Nr: 571