Fix icons for spells that only other specs have
A recent change made Cooldown Progress use the spellName instead of
the spellId to fetch information. This works better because, for
some classes the "same" spell has different spellIds depending on
spec.

But if a spec doesn't have that spell, using the spellName returns
no information. So fall back to the spell id in that case.