Treat GCD correctly
So the basic logic is:

If we have a duration > 0 and it's different from the gcd, we have
a genuine cooldown.

If we either have a duration of 0, or equal to the gcd, the ability
could have reset its cooldown.

If the old cooldown was already less than the gcd, we don't want
to reset it to 0 though.

Ticket-Nr: 613