Fix setAll for colors
In a recent commit to correct how groups and multiselection work,
I broke that by calling set(info, ..., not before), set gets passed
only the first value of in ...., but for colors 3 values need to be
passed.

So special case multiselection, and only pass the third value in
for multiselection.

The reason we need to pass in the third value for multiselection,
is that the set functions for multiselections by default only toggle
a value. If different auras in a group had different initial values,
toggling them leads to interesting behaviour.