For triggers with automaticrequired, ensure that trigger.unevent is "auto"
In the attached aura that was attached to the bug report trigger.unevent
was custom, even though there's code around line 5950 in
WeakAurasOptions.lua which sets the trigger.unevent on switching to
a automatic required trigger.

That code works. Importing the aura, switching the prototype away
and back does fix the users aura.

Ticket-Nr: 667