Fix Ignore Rune CD
We schedule a method SpellCooldownFinished() to be run once the timer
of a cooldown is finished. The moment that method needs to be run
is different for spell cooldowns that ignore the rune cd and spell
cooldowns that use the rune cd. Thus we need to use separate timers
for those.

Ticket-Nr: 820