Fix ignore rune cd for "Epidemic"
The ignore rune cd option makes wa compare the cooldown duration
returned by GetSpellCooldown with the cooldown of a rune to figure
out whether the cooldown returned by GetSpellCooldown is only
because of missing runes.

This fails for Epidemic, since the cooldown duration of that spell
is identically with the cooldown rotation for runes.

Fortunately, Epidemic is a spell with charges and thus in reality
GetSpellCharges is used. And that api never returns the rune
cooldown even if a spell is unuseable. Thus if we have charges
we don't need to check runeDuration vs duration.

Ticket-Nr: 823