Improve GetUnifiedSpellCooldown
Distinguish between spells that have a cooldown and spells that
use GetSpellCount. For spells that have cooldowns but no charges
leave charges at nil.

The Cooldown Progress (Spell) and Action Useable then interpret
convert that to 1 or 0 charge.

We can't do that in GetSpellCooldownUnified() as we the text
replacements would display show 1 charge otherwise.

Ticket-Nr: 882