Conditions: Fix check for active == false
Two separate bugs actually:
1) All checks started by checking for state and state.variable.
But that doesn't work if state.variable is false, thus change that
to state.variable ~= nil. And while doing that, remove the code
duplication for that.

2) If state doesn't exist, the code so far made the condition always
return false. But, that's incorrect for show. Make those conditions
that provide their own test also check for state. Add a new
condition test to the active condition.