Add "Hide Cooldown Text" option
Which hides the blizzard text on cooldowns. (Which is enabled
via "Show Numbers for Cooldowns".

If the user has OmniCC installed, don't show the option. As
the option would not hide OmniCC's text.

But call SetHideCountdownNumbers(true) if omnicc is installed,
as that prevents double text from appearing.

Technically, users could disable OmniCC's text and enable the
Blizzard text, but that seems a remote use case.

Ticket-Nr: 915