Make custom animation functions similar to all other functions
For some reason animation functions were saved with a "return "
prepended. For other functions we add that "return " directly before
loadstring. This meant that in various places the string needed
to be cut by 8 characters and "return " appended in others.

Somehow that code didn't work though. Instead of figuring out where
it doesn't, just remove the special casing of animation functions.

Store just the function in the variable and add "return " only
before loadstring.