Fix invalid usage of "_operator" value in trigger
If a user switches a trigger from a trigger type that has arg with
a number/longstring tpye thus using a _operator to a trigger
type that has a arg with the same name but of a different type
that is not using the "_operator" the setting for the operator
is still in the trigger data.

Fix this by only using the stored _operator setting if the type is
actually a number or a longstring.

Ticket-Nr: 979