Custom Progress options for displays that show progress
Add a way to clamp progress for the display. Values below a user
set minimum are displayed as empty progress. Progress above
a maximum are displayed as full progress.

Usefull when several displays have the same size and should move
with the same speed but their triggers have different durations.