Conditions: Fix actions not triggering twice for some conditions
For example, a condition for "Trigger 1 active" with only one trigger
only triggered once, because actions are only done when the check
changes from inactive to active.

So far, on hiding we stopped checking conditions and thus that check
was always true, and thus could only once transition from inactive
to active.

This patch adds runs another condition check on hiding, which will
deactivate all conditons and their properties.