Model Aura: Work around Blizzard bug with cameras
SetTransform is supposed to overwrite the camera that is set on
the model. But instead it apparently uses a random camera.
TheDanW promised a fix in 7.3, which unfortunately will change
the behaviour of models.

Thus, warn against using SetTransform in the option text, and
change the default for models to not use the SetTransform option.

Also reset the model transformation and others 2s after login, as
apparently the camera is different while still loading.

This fixes the aura attached to Ticket 968, and also works with
the aura in ticket 964.

Ticket-Nr: 968