Add Overlays to AuraBars/Progress Textures/Health/Power triggers
Add support for showing additional progress to AuraBars and Progress
Textures. Rewrite parts of both to enable sharing code between "normal"
progress and additional progress.

Add additional progress to the Health and Power trigger.

The Health trigger can show the Absorb and Incoming Heal as additional
progress.

The Power trigger can show the Cost of the Spell Casts

The main power is that some of the custom code people use for multiple
overlays is now much easier.

Ticket-Nr: 539