Fix interaction between animations and anchoring
The animations used to save the position/anchor on start and
restored those on finishing the animation. Even if the animation
didn't actually touch the position.

This can clash with Anchoring, as there the anchor might not be
immediately available and thus set a bit later.

Fix that, by adding 3 different orthogonal properties to regions
via RegionProperties:

- SetAnchor, which sets the anchorFrame and the anchorPoints
- SetOffset, which sets a offset
- SetOffsetAnim, which sets a extra offset for animations

Thus, if a animation ends, all it has to do is SetOffsetAnim to 0,0.

That also means that we could add a xOffset/yOffset to the conditions
now.

This is slightly backwards incompatible for custom translate functions,
they no longer get the "real" offset but the "animation" offset passed
in. And they should no longer return the "complete" offset, but the
animation offset.

Keep the old code in the animtions for the tray animation. Those still
work as before. There seems to be a bit of wonkyness with those, but
that seems to match what they did before.

Ticket-Nr: 996

squash