Fix Animations resetting Color set via Conditions
Similar idea as the previous commit.

Introduce two properties, Color and ColorAnim.

ColorAnim overrides Color if set. In the Animation set ColorAnim.
Resetting a Color animation is thus just calling ColorAnim with nil.

This kind of override/combination of properties between animations
and conditons should ideally be handled centrally instead of how
this patch implements it. This can be cleaned up at a later point.

Ticket-Nr: 945