Remove message box on using a spell id for the aura trigger
The UnitAura api takes a buff name and not a spell id, thus when
used with a spell id, we match on any buff that uses the same name.

Due to that we guided people towards using the full scan option.

Usually when people use spell ids, they want it because they want
their aura to work with all locales and not because there are
multiple buffs with shared names.

For those, the message box is confussing and probably leads people
astray.

Github-Issue: 312