Add a "In Encounter" load condition and also rework couldBeLoaded
Auras in WeakAuras can not just be loaded and unloaded but in a
"could be loaded" state. This is only used for the Options window and
decides where a aura is shown. E.g. a aura which is only loaded in
combat was still shown in the "Loaded" section and not in the "Unloaded"
section. On the other hand, auras with a encounter id, where always
shown in the unloaded section.

Fix that by adding a second load function, which only tests those
load conditions that won't dynamically change.

Github-Issue: 483