Add Spell Id to Status/Cast trigger
With the similar mechanics as Cooldown Progress. That is a spellId
matches every spell that has the same name. This takes care of the
cases where spells get replaced by different spells with exactly
the same name. Even though this makes the trigger "less exact" it
is what users mostly expect.

Github-Issue: 491