Fix showing of auras even though the options window was closed already.
We built the options via a corotinue, and as part of that show/hide auras
via that corotinue. If the corotinue is still running and the options
window is already closed, this can popup auras at an unexpected time.

Prevent the showing in PriorityShow by checking if the Options are still
open. Whether that's the best fix, or we should somehow abort at a
diffrent place, I can't really tell.

Github-Issue: 528