Fallback to region.height/width if data.height/width don't exist
WoWAce issue Nr. 1169
This is a special case for text type regions, which no longer use
width or height fields in their data. Without this, the aura data
will be mutated uncontrollably which breaks the update algorithm slightly.
Unfortunately, all data having a width and height field is a central assumption
in WeakAuras, and breaking it has many knock-on effects. This change
is the only one which does not cause a user-facing behavior change.