Add the ability to have spell/item tooltips via TSU
This was one of the few areas where the built-in triggers could
do more than TSU. Normal custom triggers can't set the tooltip though.
I think that's a okay limitation.

The TSU must respond to the "OPTIONS" event and fill in either:
- state.spellId
- state.itemId
- state.unit + state.unitBuffIndex or
- state.unit + state.unitDebuffIndex.

The values can be anything, the presence of that activates the "Show
Tooltip" option in the Display Tab.

On actually being trigger, one of those needs to be filled in with
correct data.

Github-Issue: 569