Fix nil errors due to AddMany
AddMany will Add() group parents before their children. Unfortunately this means that the regions of the children won't exist, so getrect() in RegionTypes/Group.lua will attempt to index a nil value. That's easy to guard against, but introduces a second problem, that the ticket which originally was addressed by changing getrect() will be ressurrected. So, Add() the group parents a second time, after the children have been Added. Then, the region will exist and allow the proper bounding box to be calculated.