Add conditions to normal custom and tsu triggers
There's now a additional text field for custom variables for tsu
triggers, which expects a table.

That table can list simple variables via 3 formats:
1) The full format list each variable with its type and display, e.g.
{
  var1 = {
    type = "bool",
    display = "Var 1"
  },
  var2 = {
    type = "select",
    display = "Var 2",
    values = { ... }
  }

Potential types are bool, string, select, timer, number. This is 1:1
the internal data structure that generic triggers use, so it's as
powerful as those.

2) A short hand notation for non-select variables
{
  var1 = "bool",
  var3 = "string",
}

This display name will be the same as the variable name.

3) A even shorther way for some "common" conditions, e.g.
expirationTime, duration, value, total, and stacks.

{
  expirationTime = true,
  duration = true,
  stacks = true
}

The styles can be mixed in one table.

Ticket-Nr: 984