cleanup trigger data
This is a rather large opportunity for code simplification. Currently, the first trigger and corresponding untrigger live directly in the data, while the second, third, fourth, etc triggers live in the additional_triggers field of the data. This makes interacting with multiple triggers (which is something that WeakAuras does all the time) more complex than necessary.
So, modernize the data so that all triggers are in the triggers field, which we can just iterate over much more simply.
Also, we can shunt other trigger data such as activeTriggerMode and disjunctive into the triggers field as well.