Ensure that Options Frame exists before commiting to an import job
If a user clicks on a link in chat, then the import tooltip will appear and allow them to import an aura. This is fine, but importPendingData calls WeakAuras.CloseImportExport, which assumes that the options frame has been created. So, Load WeakAurasOptions and Show the options Frame, and if that fails then back out of the import job.

All of this would be solved if we weren't abusing the tooltip like we are.
Fixes #668 on Github