Conditions: Change what we do for hidden auras
For hidden auras, ensure that no conditions get activated.

A previous fix simply prevented actions from running for hidden auras,
this goes one step further by making sure all conditions evaluate as
false for hidden auras. For example, a aura with two triggers combined
with AND and a condition on trigger 2 being true, would mark a condition
as active even if trigger 1 was not active. This prevented actions from
that condition from properly activating.

Now, the conditon won't be marked active until the aura is active and
thus any actions from that condition will fire at that moment.

This is a slight behaviour change but should be better.

Github-Issue: 677