Fix running of conditions functions for fallback states in another case
This is another old bug, that was surfaced due to how we treat hidden
auras now.

In this example, a aura was set to show the states of trigger 1, with
trigger 1 not always supplying states, thus fallback states were used.

But, the code in ApplyStateToRegion iterated over the non-existing
states, and run the conditions function to hide all states, before
creating the fallback state and reshowing the aura.

This hide+show lead to the glow reapply animation being shown.

Fix that, by moving the decision to use fallback states out of
ApplyStatesToRegion and thus using not having to do anything special
for fallback cases in ApplyStatesToRegion

Github-Issue: 721