Fix Overlay Cost of Cast interaction with SetMaxProgress
The overlay for Cost of Cast is displayed relative to the right edge
of the progress. Thus for correct display, we need to to have the
unnormalized value.

The aura in the ticket is for Soul Shards, and limits the max progress
to 1 shard. Thus at e.g. 2 shards, a overlay for a 1 shard cost spell
would be displayed to the left of the clamped progress of 1.

Essentially covering the progress bar. This commit fixes that, and
instead the overlay is displayed from the position that a progress
value of 2 would have. So essentially to the right of the bar

Since that's not exactly what the user actually wants, introduce
a Clip Overlays option, which clips any overlays to the size of aura.

Github-Issue: 724
Ticket-Nr: 1111