implement LibDeflate
From my testing, this is approximately a 80% decrease in the size of large string exports, and a further decrease in the size of addon channel transmissions.
This implementation is also backwards compatible with legacy strings, which were compressed and encoded using LibCompress, and older versions without LibDeflate will correctly detect that they cannot decode and decompress the data.
GitHub ticket #406