fix WA_IterateGroupMembers again
the previous fix caused iterations in a party to have an extra unexpected unit, like party5. For whatever reason, GetNumGroupMembers() is inconsistent with GetNumSubgroupMembers(), and GetNumSubGroupMembers() never seems to count the player itself. Since this iterator already squeezes in the player as party0, we can just use GetNumSubgroupMembers() for all party iterations