Fix not removing matchData on leaving a group
We keep a list of scanFuncs per group unit, because that seems like a
useful optimization to not check e.g. not have to filter for role/isSelf
on each check.

On leaving the group, we first updated the list of scanFuncs for party1,
and thus ScanWithFilter would immediately exit because no scanFuncs were
checking party1.

In that case we still need to cleanup the exsting match data and any
aura listening to it.

Github-Issue: 869