BuffTrigger Rewrite (#755)
Introduce a new buff/debuff tracking system

Instead of each trigger iterating over all auras, keep maps of names/spellids to triggers and use those to cut down on the amount of checking that needs to happen.

Also combine fullscan and normal auras, so that the new trigger has a superset of both modes.
Enable all of those features for group triggers, and add two more variants of group triggers for nameplates and arena units.

Use the same infrastructure also for multi-target tracking, which allows e.g. a combined view of all matches.