Try to reuse states for Auto Cloning
Previously the cloneId was simply the pointer address of the match table,
but whenever e.g. a buff with a lower index expired, the match table
might change, thus using a different cloneId.

In the aura of the bug report, using a different state was noticable,
due to having a Sound Effectr On Show.

Instad, sort all matches by unit and index and use the index as the
state key. This should in most cases make auras reuse all states and
even result in states being reused for the same aura in more cases.

It's not perfect in reusing states, but should be hopefully not
noticeable anymore.

Ticket-Nr: 1259