Remove code that checks player buffs on getting a UNIT_AURA vehicle
I'm not sure what the correct way to handle vehicles is for
bufftriggers, but just removing this code fixes a issue.

On entering a vehicle we get two UNIT_AURA events, one for vehicle and
one for pet. Both events happen in the same frame.

At the first event, UnitAura returns no auras for the player.
At the second event, UnitAura returns the auras for the player.

This is rather strange, so for now just ignore those events.

Ticket-Nr: 1270