BuffTrigger 2: Try harder to find a fallback icon
So far the fallback icon was only initialized on initially setting up
the trigger infos. Since GetSpellCooldown depends on the talents and
specs selected this might not work.

E.g. login as a Outlaw rogue and a Crimson Tempest aura won't get any
fallback.

Thus try harder to find a fallback by also looking for the spell in
LoadAura and in UpdateStateWithNoMatch.

Github-Issue: #985