fix ScanForLoads
WeakAuras.Resume calls ScanAll, which calls ReloadAll,
which calls ScanForLoads (with no args). This is the point at which
auras should be loaded initially.
But Resume occurs during the login thread, so IsLoginFinished blocks it
Therefore, split the implementation of ScanForLoads into a private func
which we call directly from ReloadAll
to skip the IsLoginFinished block