Fix isStealable check for BuffTrigger 2
On updating match data the isStealable flag was always checked due to
assigning name to it.

And for my benefit on how to reproduce that with one account + one
starter account, because it took a while until I figure out how to
easily reproduce it.

Use a high level priest and and any level 1 char. Buff the level 1 char
with Shield or Fortitude, and start a duel.

Ticket-Nr: 1276