Fix ordering of options on multi-selection
We merge the options of various different region types with a simple
union, thus if e.g.:

Icon returns
  icon.__index == 1
  position.__index == 2

And Aurabar returns
  aurabar.__index == 1
  border.__index == 2
  position.__index == 3

Then the merged list will contain
  aurabar.__index == 1
  icon.__index == 1
  border.__index == 2
  position._index either 2 or 3, depending on which gets unioned last

This means that fixMetaOrder can not figure out which of border or
position should come first. (It does that correctly only due to sorting
border/position alphabetically.)

Fix that by usign the same __index for Position and Border settings.
This is not ideal, but the correct algorithm is rather complicated.

And keep the api stable, as the stop motion plugin depends on it.