Add custom option groups
there are 2 group types, simple and array.
Simple groups have their sub options' values located directly in the group's table.
So, if you have an simple group with a key foo, and that group has a option toggle with a key bar
then you can find the value of bar at aura_env.config.foo.bar.

Array groups have copies of their sub options' values in an array of 0 or more subtables of the group's table.
This allows the user to make multiple similar entries with similar data
(quite handy for e.g. setting up an interrupt rotation).
If you have an array with key foo, and a sub option in that array with key bar,
then you may find the values of bar at aura_env.config.foo[index].bar,
where index is an integer which represents the...index of the entry you are searching.
Entries into an array always constitute a contiguous range from 1 to #aura_env.config[key], where key is the key of an array group.