Fix a few issues around MSQ integration
We need to fetch a new group if our id changed. Also use a id that is
only unique per id, instead of reusing frameId.

FrameId is unique over all frames, but that hurts the masque
integration in that it creates new groups per icon in a cloning trigger.

This used to work because the frameId algorithm didn't acutally assign
unique ids until recently.

Github-Issue: #1111