Fix Cooldown setting not being enabled for some settings
The aura in this case is a totem trigger, which a enabled but hidden
inverse setting.

The generic trigger's CanHaveDuration controls whether the Cooldown
setting is enabled or not.

For a status/event trigger, it checks if the prototypes contains a
durationFunc or the canHaveDuration bool, but also checks if inverse is
enabled. For most triggers, enabling inverse means there's no timing
information. (E.g. not on cooldown.)

Thus the bug is that CanHaveDuration didn't take into account that the
inverse setting is hidden.

But, I don't think handling inverse magically in CanHaveDuration is a
good idea, so instead of fixing it by checking that the inverse setting
is hiden, simply don't check for it at all.

That means the Cooldown checkbox will be enabled even if doesn't do
anything in a few more cases, but makes CanHaveDuration depend on less
magic.

Also refactor CanHaveDuration to be easier to read.

Github-Issue: #1120