Partily revert 7f62571f434c, Masque integration fixes
Masque doesn't unskin buttons correctly.

We reuse buttons e.g. two buff triggers checking for buffs on
focus/target will reuse the same button. If a button was already in a
group and we change the masqueId and thus want to add the button to a
new group, but the new group is disabled, masque applies a "Blizzard"
skin.

So revert that for now, and report a bug for masque

Fixes: #1127