Fix collapse setting on multi selection
The basic problem is how the display tab works on multi-selection. If we
have a multi selection we generate a ace options table for each
individual aura and one for the group.

The group one gets all functions removed and replaced by *all methods.

So, if we have two auras with different collpase states, their image
function will return different images, and thus the imageAll can't find
a common image and will return no image.

This commit "fixes" that by not making the collpase state a per
aura setting, but one shared by all auras.

Github-Issue: #1088