Optimize Unit Change events for units that have no events
By creating a separate frame that can quickly check if a unit's guid did
change and sends a fake event for just that unit to all interested
auras.

Creating such a fake event fits the current system, but feels like it
might be worth refactoring how the events to aura mapping works.

This should fix:
Fixes: #1210