Removing the "Hide: Custom" option from most triggers
Mostly for cases where it appears to be broken.

But also from:
Health + Power, where it could technically be used, but I find it very
unlikely that anyone is using it. And the interface is very confussion.

Keep it for the Chat Message trigger, as I know from past bug reports
that it is indeed used.

Fixes: 897