assume that get(...) == nil is equivalent to false
This only fixes the symptom, not the underlying problem.
The code which generates the combined info for triggers is flawed.
And assumes that get(...) always returns a non-nil value.
We could embrace that assumption and enforce a non-nil value in all of our Ace get methods
Or, we could adjust this code path to not makes this flawed assumption
Fixes #1335