Ensure that trigger_events is a proper array and doesn't skip any index
E.g. UNIT_POWER:player UNIT_PET would result in
trigger_events = { nil, UNIT_PET }.

And while no code seems to rely on that being a array, it's bad that it
looks like a array in all but a few corner cases, so change it before it
becomes a problem.