Fix Models that should show on login
Models that aren't shown are cleared of their models automatically by
wow unless SetKeepModelOnHide is set. Before a1cc2b4, we used to reset
the model on various events that cleared models, such as movies and
worldmaps. Since that commit we use SetKeepModelOnHide, but it turns out
models that are shown immediately on login don't show up.

So after login, run PreShow once on each Model that should be visible.

Fixes: #1387