Fix Progress Bars size calculation in SetValue
The Progress Bars SetValue gets the size of the total bar via
GetWidth/GetHeight. But due to Blizzard's layout engine being too lazy
to calculate these values both functions return 0.

Even though we set the width of the whole region to the correct value
before we ask for the width of the bar.

Thus, instead of relying on Blizzard's api to layout the bar, calculate
the width of the bar from our data and use that one

Fixes: #1384