Make Auras attached to nameplates work inside Dynamic groups
On auras attached to nameplates, we can't call GetPoint().
The animation code doesn't actually need to call GetPoint() if the to be
animated region provides a SetOffsetAnim function.

Thus provide such a function for controlPoints.

Also we can't call Get(Left/Right/Top/Bottom), so wrap these in xpcall
and assume 0 for their position. This therotically breaks background
positioning, but that's incompatible with the hacky attaching anyway.

Fixes: 1373