improve Restoration functionality
Just Add()ing things doesn't update options. So do all of that.
This is mostly a clone of ConvertDisplay, there may be some improvements to be made.
and add a few TODO comments about what systems need to be migrated
before the new RestoreDisplay function can use only the uid