BT2: Add support for filtering by class
* Rewrite how we calculate max count for groups. Count it per
triggerInfo on updating to which group units a triggerInfo applies.
This is much easier than precalculating every combination of
class/role/group type.

* The options code for the buff trigger is different from the generic
trigger, it's not trivial to have the same one+multiple selection
widget that the generic trigger has. Thus simply use a multi selection.

This implements another part of: #1756