Fix UnitIsUnit check for Unit Characteristics
This is messy. The design of the unit triggers is, that they have one
state for each unit id that is relevant, and that they are only
interested in events that affect that unit. And thus events are
forwarded to the state that corresponds to the unit.

Unfortunately, the UnitIsUnit check breaks that. With that, a e.g. focus
unit with a UnitIsUnit check for target is interested in
UNIT_CHANGED_target events.

Fix this by:
a) Removing the UnitIsUnit check for group triggers. This is too much
work to fix, it would need to forward that event to each state instead
of just one state.

b) Hardcode the unit we pass into the trigger function for non group
triggers.

The UnitIsUnit check really doesn't fit into the design.

Fixes: #2152