Tweak trigger/conditions page by defaulting to collapsed if necessary
That with more than 1 trigger or 2 conditions, collapse them all.

Currently distinguishing the triggers while collapsed is impossible,
that is something to fix soon.

Also fix a few bugs around moving triggers/conditions around to keep the
collapsed state.

And move the Add trigger button to the bottom to mirror how the page
works for sub elements and conditions

Fixes: #2479
Fixes: #2399