Refactor Delete/Rename code in Options
The Options have to do a bit more cleanup on renaming/delete. This
changes how this cleanup code is run.

The WeakAuras.Rename/Delete code now calls a callback to which the
options responds by doing its own cleanup.

This has two benefits:
a) This moves us closer to a world where individual systems can be
loosely coupled via a few callbacks, similarly to what I did for the
AuraWarnings system, which has a small interface of just two methods.

b) This moves us a tiny bit closer to Undo/Redo