Hack barmodels with the same PreShowModels hack that Models have
As always models are absolutely mysterious. In this bug report a icon
has both a barmodel and a glow applied to it. If both are visible, on
reload the model won't show up. If the border is not visible then it
does. No idea how the glow affects the model.

The existing hack for models when applied to barmodels fixes the issue.
So apply it and say goodbye to any hope that barmodels could live with
less hacks.

Fixes: #2519