Rename a few anchor options
For a Progress Bar, "Bar" is unfortunately ambigious, because it's
unclear whether the icon is included or not.

Rename it to "Full Bar" or "Background" depending on what is meant.

Background is not a good name though, since it will lead to users
wondering whether a text anchored to the background is behind the
foreground.

Fixes: #2594