Introduce "elapsedTime" as a conditionType
And modify a few triggers to make use of it:

* The buff trigger provides:
  * apply time
  * apply/refresh time
  * stack gain time
  * stack lost time

* Cooldown Progress (Spell) + Action Usable
  * ready time
  * charge gain time
  * charge lost time

This feature will shine once animations are controllable via
conditions, as that should make it possible to show a animation for a
few seconds on any of those events.

This also works for firing actions, e.g. playing a sound, though it's
a bit akward. Though that's nothing new for actions in conditions, and
does mirror how remaining time checks work.

Also missing is any way to show a text for the time since the event.

Fixes: #1357
Fixes: #1447