Add a raw instance type load/trigger option
This has a few pros:
* Automatically updates itself whenever Blizzard adds a new type of
content.
* Adds a Torgast load option
* Would have been potentially useful in the past for e.g. Warfronts or
Visions or Islands.

And a few cons:
* There are quite a bunch of legacy types in there, which are probably
no longer used and thus the list is long and ugly.
* The names are far from self-explanatory.

I added a magic system to disambiguate names, but we might need to
tweak this regularly or even overwrite what the api returns and I don't
even know where some of the ids are used.

Nevertheless this is probably useful enough.

Fixes: #2858