Make Time Formats more flexible
By making the dynamic threshold configurable.

And by using Blizzard's time formatting functions, as proposed by
Elvador in #2173. This patch offers basically the same functionality,
though quite differently presented to the user.

There's now a top-level option to choose between:
* Built-in:            63:42   3:07    10    2.4
* Old Blizzard         2h      3m 7s   10s   2.4
* Modern Blizzard      1h 3m   3m 7s   10s   2.4

All formats support (and default to) increased precision below a
threshold. The default threshold remains at 3s.

The threshold setting is also the off-switch for the additional
precision. Setting it to 0, disables that behaviour. This isn't the most
obvious user interface, but creating a special control for it is too
much work and a additional checkbox destroys the layout.

Fixes: #2707