Fix nil errors in dynamic conditions
The first issue, was that for unit triggers we automatically create
the state, which could result in a state that isn't shown and is
otherwise empty being left in activeStates.

Also fix one missing Start/Stop Profile and a missing environment
activation.

And if the region is hiding, don't check the recheck time

Fixes: #2939