Fix creating Model Aura from Template for a Ability
We do have a list of templates that apply per template category, that is
e.g. "Show on Cooldown" or "Show Cooldown and Buff". These templates use
conditions and thus can't be used with all region types. They are thus
filtered.

For model auras we ended up without any template, leading to a error.

Fix that by including a fallback template that doesn't use conditions,
and use these if no other templates work.

Fixes the last issue mentioned in:
Fixes: #2956