Cast trigger: Show a "Exact Spell match" option for the spellId check
Technically the spell name is now entirely redundant with the spell id
check, but I don't know what to do with the spell name check.

The options are to:

a) Try to modernize it by calling GetSpellInfo on the name and putting
it into the spell id check.

This does "break" importing "old" auras from a different locale.
In that previously you'd see the spell name of the wrong locale, whereas
after such a modernization the spell name would simply be gone.

b) Add a compability option for the spell name check.

Fixes: #3073