Deprecate some old sound files
For notification sounds, which is the main purpose of sounds in
WeakAuras, we want short, front-loaded, punchy sounds.

We did inherit a lot of sounds from PowerAuras that don't really fit
into that criteria, so mark them from deprecation.

The benefit is a higher quality selection of sounds.

Let's see how many people complain, we'll probably end up keeeping a few
of these.