Fix Toggle List Default layout
We use a multiselect widget, which creates a InlineGroup. That
InlineGroup uses up some vertical space, so setting the width of each
checkbox to WeakAuras.normalWidth can lead to a undesirable layout.

Decreasing the width to 90% of that should not be noticeable to the
user, while guranteering the intended layout.

Fixes: #3371