Let timestamp differ slightly for extra attacks
An arms warrior's sword specialization talent will trigger the event
SPELL_EXTRA_ATTACKS with a SWING_DAMAGE/SWING_MISSED immediately
after. The swing timer should not be reset by the extra attack, but
the check requires the exact same timestamp for SPELL_EXTRA_ATTACKS
and the following SWING_DAMAGE/SWING_MISSED.

The events are timestamped separately, thus the timestamp does not
strictly have to be the same. Allow a small discrepancy between the
timestamps to not reset the swing timer on mid-swing sword
specialization procs unnecessarily.