Fix Modernize calling GetProperties lua error
Calling GetProperties in Modernize was a bad idea, because we obviously
run Modernize before we add. But GetProperties does depend on data that
is generated on Add. That is the overlay information.

Fortunately we don't all properties, all we need are the properties for
sub elements, and those don't depend on Add.

While the lua error log is pretty obvious that this is the problem, I
can't actually test this against real settings, because none of the
affected people actually properly reported the bug by including their
settings file.

Fixes: #3436