Fix inconsistency of text replacements
We have a shortcut path for 2 letter strings, e.g. %p, %t. Because those
are very common, and we don't need to parse it.

Unfortunately this code path used a different logic for deciding what to
do with %c, if the state contains a .c member.

We additionaly had logic to hide the formatter for %c replacements. But
that was incorrect becasue %c could refer to state.c. (In the not
shortcut path.)

Fix that by removing the special casing of %c for the options, and since
we now show the options make them actually work for custom texts.

That's a minor feature, which got requested a few times.

Fixes: #3561