BT2: Fix stale/wrong information in multi-tracking states
The unit property was only set on constructing the state, and not
updated if the mapping between guids and units changed. While it is
possible to write code that updates all states, that seems like
unnecessary complexity, since apparently no one depends on unit.

The GUID property was wrong because it was updated from the stale unit
information. Removing the unit property, makes the GUID stay correct.

Fixes: #3711