Fixup broken settings after a bug in update.lua
Unfortunately there was a bug in update.lua that lead to auras creating
a circle.

And even more unfortunately there was a second bug in the code that
should have printed a error message. So the bug didn't get properly
diagnosed.

I did analyze two broken settings, and while for the first detection of
which parent is wrong was easy, the second case requires a hardcoded