Tweak decompression error message
We switched to a different compression method 2.5y ago, so at this time
there should be very few auras that use the old compression.

On the other hand, we have plenty of people trying to import plater or
elvui profiles into WeakAuras.

Give them an error message that puts them on the right track.