Make IsSpellKnown* checks check the base spell more correctly
IsSpellKnownOrOverridesKnown returns true for e.g. Defile even if Defile
is not talented, because Defile's base spell is Death and Decay which is
always known.

Thus for that example, additionally check whether Death and Decay is
actually overriden by Defile.

Also remove a old workaround for Void Eruption, since this logic should
apply to it as well.

Fixes: #4005