Texture/Progress Texture: Improve atlas compatibility
For Textures, use texture:SetRotation instead of a tex coord based
rotation, as that's identically.

That obsoletes the "full rotate" toggle, but we keep it as a "Legacy Zoom
Out" to keep backwards compatibility.

For Progress Textures using texture:SetRotation instead of tex coord
rotation would come with several behaviour changes, so instead offer it
as an additional option called "Aura Rotation".

Keep the tex coord manipulation unchanged, the aura rotation is on top
of the tex rotation.

Atlas textures are incompatible with the tex coord manipulation we do,
but don't hide the options.

I tried masking the outside of atlas textures, but due to a wow bug:
https://github.com/Stanzilla/WoWUIBugs/issues/353
this work rests in the atlas branch on my github.

Fixes: #4027