Cooldown Tracking: Don't react within 2s of PLAYER_ENTERING_WORLD
Apparently if the player is specced into talents, on each
PLAYER_LEAVING_WORLD/PLAYER_ENTERING_WORLD, that is each portal the
player unlearns and relearns talents.

If a talent grants a second charge, this can lead to the
SPELL_CHARGES_CHANGED event.

So if it's been <2s since the last PEW, don't do anything except
installing a OnUpdate handler.

After 2s have passed the OnUpdate handler updates the spell cooldowns
and uninstalls the OnUpdate handler.

Thus for 2s after PEW the information that WA provides might be
outdated.

Fixes: #4061