AuraBar: Fix regression in textures after gradient introduction
In ffae60d6f9e37266b, instead of adjusting the foreground size and
texture coordinates, we switched to having a frame that clips the
foreground texture, and the code for adjusting the texture coordinates
was removed.

The texture coordinates code actually had 2 jobs, the adjustment to the
foreground size and the rotation of the texture for vertical progress
bars.

Reintroduce the texture coord setting, but only do it on orientation
changes.

Fixes: #4087