Talents: Rework checking for active talents
The old method unfortunately didn't work for a few selected spells, so
we need to check via the talent api.

But that api doesn't make it easy to check if a given talent is active
or not.

So build a cache that contains the information per spellId on how many
ranks are selected.

Fixes: #4197