Improve TimerTick/Frame system
For TimerTick sub elements and the region itself now subscribe to the
TimerTick event. The first to subscribe/last to unsubscribe causes the
region to install/remove an OnUpdate handler.

The high priority flag gurantees that the region gets the TimerTick
first, even though I'm unsure whether that's actually required, but it
keeps the same behaviour.

A better system for ordering subscribers will be needed at some point.

For FrameTicks the system has now two layers. The base is
Private.FrameTick, which if there are any subscribers installs an
OnUpdate handler.

Each region subscribes to Private.FrameTick if there are subscribers to
their FrameTick.

Also factor out the code into a separate file.