Weapon Enchant: Fix enchant trigger not parsing name correctly
Tooltip parsing is complicated. In this case while the text appears on
the tooltip to be "Enchant (1 hour)" and our pattern does match that.

The actual text is actually "enchant (1 |4hour:hours;), which our
pattern does not match. Fun fact, printing such a string obviously
automatically strips the ui escape sequence.

So tweak it to only require ONE non-number character after all numbers.

This is to distinguish it from the pvp item line.

And that, kids, is why we don't do tooltip parsing.

Fixes: #4216