BuffTrigger2: Check that matchDataByTrigger contains what we expect
The matchData.auras and matchDataByTrigger are supposed to be in sync
with each other, but apparently they can get out of sync.

I have stared at the code for some hours and can't find any way for
these data structures to get out of sync.

Nevertheless guard against that possibility.

Fixes: #4373