Fix issue with Smooth Progress, Min/Max clamps and overlays
If both smooth progress and min/max progress settings were enabled,
overlays could be positioned inorrectly because the smooth progress bar
clamps the value.

So e.g. a true progress value of 1.2 with an backwards overlay with
width 0.5, should start at 0.7 and extend to 1.2.

If the progress is clamped to 0..1, then that doesn't work because the
smoothed progress value will at max be 1.

Fixes: #4400