In ActivateEnvironment for tsu variables, don't create the region
The ActivateAuraEnvironment for tsu variables is run in "configMode",
which is supposed to not run the init function.

In the linked bug report, due to a EnsureRegion call the region was
created and due to a custom anchor function then run a full
ActivateAuraEnvironment.

Fix that by not creating a region.

Fixes: #4475