BT2: Add InRange check
Noteable blizzard only sends those events if there's a RegisterUnitEvent
that for a specific unit. They probably are optimizing the checking to
only Register the event if a buff trigger is loaded that requires it.

Also we dance a bit to ensure that we don't register for pets unless
necessary.

Fixes: #4542